pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "Anime-Recommender"
version = "0.1"
authors = [
    { name = "Aman" }
]
dependencies = [
    "pandas",
    "numpy",
]

[project.optional-dependencies]
deployment = [
    "flask",
    "requests",
    "joblib",
    "gunicorn"
]
development = [
    "google-cloud-storage",
    "pyyaml",
    "tqdm",
    "tensorflow",
    "scikit-learn",
    "comet-ml",
    "dvc[dvc-gs]"
]

[tool.setuptools.packages.find]
where = ["."]